Calculate Distribution Overlap

calc_drawsoverlap() calculates the generalized distribution overlap according to Weitzman (1970) for a draws_matrix.

Usage

calc_drawsoverlap(draws_mat)

Source

Weitzman, M. S. (1970). Measures of overlap of income distribution of white and negro families in the United States (Technical paper No. 22). United States Bureau of the Census. https://hdl.handle.net/2027/mdp.39015085502204.

Arguments

draws_mat

A draws_matrix

Value

A n x n numeric matrix

Details

A draws_matrix is a numeric matrix in which the rows are posterior draws and the columns are variables.

References

Al-Saleh, M. F., & Samawi, H. M. (2007). Interference on overlapping coefficients in two exponential populations. Journal of Modern Applied Statistical Methods, 6(2), 503–516. https://doi.org/10.22237/jmasm/1193890440.