Pooling Factors for a Multilevel Model

calc_poolfac() calculates the pooling factors according to Gelman and Pardoe (2006) from a CmdStanMCMC object

Usage

calc_poolstats(y, draws_mat, pars = c("gamma", "delta", "y"))

Source

Gelman, A., & Pardoe, I. (2006). Bayesian measures of explained variance and pooling in multilevel (hierarchical) models. Technometrics, 48 (2), 241–251. https://doi.org/10.1198/004017005000000517.

Arguments

y

A numeric vector with the response variable

draws_mat

A draws_matrix with the relevant pooling parameters

pars

A character vector with the names of the relevant residuals of the multilevel model

Value

A named vector containing the pooling factors

Details

A draws_matrix is a numeric matrix in which the rows are posterior draws and the columns are variables.