Standard Error of Measurement and Confidence Intervals

calc_semci() returns the standard error of measurement from a given reliability and the corresponding standard deviation. The confidence intervals are calculated using a normal or a student t distribution, or according to the Tschebyschow inequality for an unknown distribution.

Usage

calc_semci(data, sd, rel, se0 = 2.8, p_ci = NA, dist = "none", ...)

Arguments

data

A numeric vector containing the measured values

sd

A numeric vector of standard deviations of measurement scales

rel

A numeric vector of reliability coefficients of measurement scales

se0

The number of standard errors from 0 which are regarded as the critical value for the confidence interval which defaults to se0 = 2.8

p_ci

The confidence niveau in percent which defaults to `p_ci = NA'

dist

One of "normal", "t", or "none"

...

The number of degrees of freedom for the student t distribution

Value

A data.frame containing the inputs, the standard error of measurement, and the lower and upper bounds of the confidence interval

Details

One of the arguments se0 or p_ci must be set to a value. The other must be set to NA.