Content Validity of Tests for Psychological Sense of Community

The rated content validity of the items of 4 different tests in specific community contexts.

Usage

conval_data

Format

contentvalidity

A data frame with 5,609 rows and 79 columns.

Variables

person

Person number from conval_demo

comm

The specific community from conval_comms for which the test items were rated

value

The grade to which the community is a community of shared values

purpose

The grade to which the community is a community of purpose

value_x_purpose

Interaction variable of value and purpose

chavis_[1-25]

The grade to which the test item [1-25] from Chavis et al. (2008) is rated necessary to characterize sense of community

omoto_[1-18]

The grade to which the test item [1-18] from Omoto and Snyder (2010) is rated necessary to characterize sense of community

jason_[1-9]

The grade to which the test item [1-9] from Jason et al. (2015) is rated necessary to characterize sense of community

halamova_[1-22]

The grade to which the test item [1-22] from Halamová et al. (2018) is rated necessary to characterize sense of community

Source

Primary data

References

Chavis, D. M., Lee, K. S., & Acosta, J. D. (2008). The Sense of Community Index (SCI) revised: The reliability and validity of the SCI-2. Paper presented at the 2nd International Community Psychology Conference. Lisboa, Portugal. Retrieved October 17, 2018, from https://senseofcommunity.com.

Halamová, J. Kanovsky, M., & Naništová, E. (2018). Development and psychometric analysis of the sense of community descriptors scale. Psychological Intervention, 27(1), 44-55. https://doi.org/10.5093/pi2018a8.

Jason, L. A., Stevens, E., & Ram, D. (2015). Development of a three-factor psychological sense of community scale. Journal of Community Psychology, 43(8), 973-985. https://doi.org/10.1002/jcop.21726.

Omoto, A. M., & Snyder, M. (2010). Influences of psychological sense of community on voluntary helping and prosocial action. In S. Stürmer & M. Snyder (Eds.), The psychology of prosocial behavior: Group processes, intergroup relations, and helping (pp. 223-243). Wiley-Blackwell. https://doi.org/10.1002/9781444307948.ch12.