Plot Data for Cluster Probabilities

make_lambdapointrange() prepares the data in the format for the function plot_lambdapointrange().

Usage

make_lambdapointrange(draws_mat_list, labels, probs = 0.87, tikzdevice = FALSE)

Arguments

draws_mat_list

A list of draws_matrix objects containing the lambdas

labels

A character vector containing the names of the graph facets

probs

A numeric value representing the probability for the highest density region which defaults to probs = 0.87

tikzdevice

Logical indicator for preparing the text for tikzDevice which defaults to tikzdevice = FALSE

Value

A data frame in the format for using the corresponding plot_dp_prcstats() function

Details

A draws_matrix is a numeric matrix in which the rows are posterior draws and the columns are variables.