Sample Characteristics

The characteristics of the samples for the studies.

Usage

comms_demo

conval_demo

Format

comms_demo

A data frame with 158 rows and 7 columns.

conval_demo

A data frame with 1,465 rows and 7 columns.

Variables

person

Person number

gender

Gender coded as 1 = Male, 2 = Female, 3 = Divers

age

Age in years

state

One of the 16 German states the person is living in alphabetical order or 17 = living outside from Germany

studstatus

Student status used in the FernUniversität official statistics (Berichtswesen, 2022, 2023)

noexpert

Formal theoretical knowledge of a theory of psychological sense of community coded as 0 = Knowledge, 1 = No knowledge

numcomm

The number of communities the person mentioned

Source

Primary data

Berichtswesen. (2022, March 25). Studierendenstatistik für Sommersemester 2018 [Student statistics for summer semester 2018]. FernUniversität in Hagen. https://www.fernuni-hagen.de/uniintern/organisation/statistik/semesterstatistik/sose2018.shtml.

Berichtswesen. (2023, March 31). Studierendenstatistik für Wintersemester 2019/20 [Student statistics for winter semester 2019/20]. FernUniversität in Hagen. https://www.fernuni-hagen.de/uniintern/organisation/statistik/semesterstatistik/wise2019-20.shtml.