Procrustes Transformation of MDS Configuration

transform_procrustes() calculates the transformation of a multidimensional scaling configuration Y to a target configuration X.

Usage

transform_procrustes(x_mat, y_mat, dilation = FALSE)

Arguments

x_mat

A numeric matrix representing the MDS target configuration

y_mat

A numeric matrix representing the MDS configuration to be compared

dilation

A logical indicator for implementing optional scaling which defaults to dilation = FALSE

Value

A numeric matrix representing the transformed MDS configuration